moved to src/main/java, changed onPlaceholderRequest to onRequest for 2.10.7 compliance.
